refactor: API 타입을 api 폴더로 이동
